Update Main.java some general updates
Remove the static from each method, consequently  I had to add a attribute scanner, so that every method in the class can call the scanner. 

I also moved the mainMenu() from the main(String[] args) because the (String[] args) is meant to initiate the program, by having the method mainMenu() within it, it would just cause the program to start slower. 

Also because we are adding submenus, is best to keep the menus outside the (String[] args).