Rework on Main.java
Main.java now has a proper menu, with sub menus for either doctors or patients. 

Move every CSV related method to another class DataHandler this way the data is better secured and the main.java is just about running the inputs for the user.

Note that is just missing some javadoc, and some of the println() are in english.