Create DataHandler.java to manage every CSV
So for security concerns and efficiency, I moved every method in Main that was about just dealing with CSV files and moved them here. 

I also changed some of the type of methods to work better with the methods in Main. 

Most importantly, I created three distinct csv files for each class, this way the data for clinica, paciente, or doctor are not in the same file. This is mainly due to security concerns.

Note: some methods are missing javadoc.